División de código completada: Compila
