+++ Arduino files/MaquiRubi_main/controlSerial.h
@@ -1,11 +1,15 @@
+#ifndef __controlSerial_h__
+#define __controlSerial_h__
+#include <Arduino.h>
+
 #define MESSAGE_LENGTH 100
 
-
-char mensaje[MESSAGE_LENGTH];
-char test_mensaje[MESSAGE_LENGTH]="FLDFLDFLDfldFLdfLD";
+extern char mensaje[MESSAGE_LENGTH];
+extern char test_mensaje[MESSAGE_LENGTH];
 
 enum State {INICIO=0,FOTOS,IA,ENVIO_DATOS,EJECUTANDO};
 
 void getMensaje(char* msg);
-SentidoGiro getSentidoGiro(char instruccion);
 int saveData(char message[], int MAX_LENGTH);
+
+#endif
